requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autocorrect==2.6.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
config==0.5.1
demoji==1.1.0
docopt==0.6.2
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
nltk==3.8.1
num2words==0.5.13
openai==0.28.1
packaging==23.2
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
